# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  firs_models.FirsWebhookPayload:
    properties:
      irn:
        type: string
      message:
        type: string
    required:
    - irn
    - message
    type: object
...
  firs_models.UpdateInvoice:
    properties:
      payment_status:
        type: string
      reference:
        type: string
    required:
    - payment_status
    type: object
...
  models.Response:
    properties:
      data: {}
      error: {}
      extra: {}
      message:
        type: string
      name:
        type: string
      pagination: {}
      status:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
  description: This is the e-invoicing service API documentation.
  termsOfService: http://swagger.io/terms/
  title: Gention E-invoicing Service API
  version: "1.0"
paths:
...
  /invoice/business/{business_id}:
    get:
      consumes:
      - application/json
      description: Returns a list of invoices with minimal details for a business
      parameters:
      - description: Business ID
        format: uuid
        in: path
        name: business_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get all invoices by business ID
      tags:
      - Internal Invoice
  /invoice/business/{business_id}/{invoice_id}:
    delete:
      consumes:
      - application/json
      description: Deletes an invoice by business_id and invoice_id
      parameters:
      - description: Business ID
        format: uuid
        in: path
        name: business_id
        required: true
        type: string
      - description: Invoice ID
        format: uuid
        in: path
        name: invoice_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Delete Invoice
      tags:
      - Internal Invoice
    get:
      consumes:
      - application/json
      description: Returns full invoice details by business ID and invoice ID
      parameters:
      - description: Business ID
        format: uuid
        in: path
        name: business_id
        required: true
        type: string
      - description: Invoice ID
        format: uuid
        in: path
        name: invoice_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get one invoice details
      tags:
      - Internal Invoice
  /invoice/confirm/{irn}:
    get:
      consumes:
      - application/json
      description: Confirms an invoice with IRN.
      parameters:
      - description: Invoice Reference Number (IRN)
        in: path
        name: irn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice confirmed successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Confirm Invoice
      tags:
      - Invoice
  /invoice/create:
    post:
      consumes:
      - multipart/form-data
      description: Upload a JSON invoice file and store it in DB
      parameters:
      - description: Invoice JSON File
        in: formData
        name: file
        required: true
        type: file
      - description: Business ID
        in: formData
        name: business_id
        required: true
        type: string
      - description: Invoice Number
        in: formData
        name: invoice_number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice created successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Create a new Invoice
      tags:
      - Internal Invoice
  /invoice/download/{irn}:
    get:
      consumes:
      - application/json
      description: Downloads an invoice from FIRS using the IRN.
      parameters:
      - description: Invoice Reference Number (IRN)
        in: path
        name: irn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice downloaded successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Download Invoice
      tags:
      - Invoice
...
  /invoice/update/{irn}:
    patch:
      consumes:
      - application/json
      description: Updates an existing invoice using the IRN.
      parameters:
      - description: Invoice Reference Number (IRN)
        in: path
        name: irn
        required: true
        type: string
      - description: Update Invoice Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/firs_models.UpdateInvoice'
      produces:
      - application/json
      responses:
        "200":
          description: Invoice updated successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Update Invoice
      tags:
      - Invoice
...
  /webhook/firs:
    post:
      consumes:
      - application/json
      description: Receives webhook events from FIRS (e.g., IRN status updates).
      parameters:
      - description: FIRS Webhook Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/firs_models.FirsWebhookPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Webhook processed successfully
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad request - invalid body or processing error
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/models.Response'
      summary: FIRS Webhook Receiver
      tags:
      - Webhooks
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"